# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: FUNHUB
  description: ''
  version: 1.0.0
servers:
  -
    url: 'https://funhub-mobile-backend.test'
paths:
...
  /api/v1/products/checkout:
    post:
      summary: 'Post Checkout'
      operationId: postCheckout
      description: ''
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {USER_TOKEN_KEY}'
          schema:
            type: string
      responses:
        200:
          description: fiat_payment_mode
          content:
            text/plain:
              schema:
                type: string
                example: "{\n\"message\": \"Redirect to gateway\",\n\"gateway_data\": [\n     'url' => $this->url .'/payment/eCommerce',\n      'secureHash' => $this->generateHashForRequest($this->mid, $invoice_no, $amount),\n      'mid' => $this->mid,\n      'invno' => $invoice_no,\n      'capture_amt' => $amount,\n      'desc' => $desc,\n      'postURL' => $redirectUrl,\n      'phone' => $phoneNo,\n      'email' => $email,\n      'param' => $param,\n      'authorize' => 'authorize'\n  ];\n}"
      tags:
        - Product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                product_id:
                  type: integer
                  description: 'Product ID.'
                  example: 1
                payment_method:
                  type: string
                  description: 'Payment Method.'
                  example: fiat
                fiat_payment_method:
                  type: string
                  description: 'required_if:payment_method,fiat Payment Method.'
                  example: fpx/card
                quantity:
                  type: integer
                  description: Quantity.
                  example: 1
              required:
                - product_id
                - payment_method
                - quantity
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []